Moved old commands to separate dir
